upd applyStyles module
