+++ packages/react/src/lib/applyStyles/index.ts
@@ -1,0 +1,5 @@
+import applyStyles from './applyStyles'
+
+export default applyStyles
+
+export * from './applyStyles'
